OPIK-329 move AppMetadataService to infra
