+++ apps/opik-backend/src/main/java/com/comet/opik/infrastructure/health/IsAliveResource.java
@@ -2,7 +2,7 @@ package com.comet.opik.infrastructure.health;
 
 import com.codahale.metrics.health.HealthCheck;
 import com.codahale.metrics.health.HealthCheckRegistry;
-import com.comet.opik.domain.AppMetadataService;
+import com.comet.opik.infrastructure.AppMetadataService;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
